backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic[email]>=2.0.0
httpx>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
beanie>=1.25.0
motor>=3.3.2
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0